pyproject.toml
[project]
name = "pallete-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "mcp[cli]>=1.6.0",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
